♻️ Refactor serialize functions to one class
- Move setup serialization functions to one class
- Refactor tests.
